+++ commercetools-models/src/main/java/io/sphere/sdk/taxcategories/SubRateImpl.java
@@ -1,18 +1,16 @@
 package io.sphere.sdk.taxcategories;
 
-import java.math.BigDecimal;
-
 final class SubRateImpl implements SubRate {
     private final String name;
-    private final BigDecimal amount;
+    private final Double amount;
 
-    SubRateImpl(final String name, final BigDecimal amount) {
+    SubRateImpl(final String name, final Double amount) {
         this.name = name;
         this.amount = amount;
     }
 
     @Override
-    public BigDecimal getAmount() {
+    public Double getAmount() {
         return amount;
     }
